# strava/aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Strava API v3",
    "description" : "The [Swagger Playground](https://developers.strava.com/playground) is the easiest way to familiarize yourself with the Strava API by submitting HTTP requests and observing the responses before you write any client code. It will show what a response will look like with different endpoints depending on the authorization scope you receive from your athletes. To use the Playground, go to https://www.strava.com/settings/api and change your “Authorization Callback Domain” to developers.strava.com. Please note, we only support Swagger 2.0. There is a known issue where you can only select one scope at a time. For more information, please check the section “client code” at https://developers.strava.com/docs.",
    "version" : "3.0.0"
  },
  "servers" : [ {
    "url" : "https://www.strava.com/api/v3"
  } ],
  "security" : [ {
    "strava_oauth" : [ "public" ]
  } ],
  "paths" : {
    "/athletes/{id}/stats" : {
      "get" : {
        "tags" : [ "Athletes" ],
        "summary" : "Get Athlete Stats",
        "description" : "Returns the activity stats of an athlete. Only includes data from activities set to Everyone visibilty.",
        "operationId" : "getStats",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "The identifier of the athlete. Must match the authenticated athlete.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Activity stats of the athlete.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ActivityStats"
                }
              }
            }
          },
          "default" : {
            "description" : "Unexpected error.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Fault"
                }
              }
            }
          }
        }
      }
    },
    "/athlete" : {
      "get" : {
        "tags" : [ "Athletes" ],
        "summary" : "Get Authenticated Athlete",
        "description" : "Returns the currently authenticated athlete. Tokens with profile:read_all scope will receive a detailed athlete representation; all others will receive a summary representation.",
        "operationId" : "getLoggedInAthlete",
        "responses" : {
          "200" : {
            "description" : "Profile information for the authenticated athlete.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DetailedAthlete"
                },
                "example" : {
                  "id" : 1234567890987654321,
                  "username" : "marianne_t",
                  "resource_state" : 3,
                  "firstname" : "Marianne",
                  "lastname" : "Teutenberg",
                  "city" : "San Francisco",
                  "state" : "CA",
                  "country" : "US",
                  "sex" : "F",
                  "premium" : true,
                  "created_at" : "2017-11-14T02:30:05Z",
                  "updated_at" : "2018-02-06T19:32:20Z",
                  "badge_type_id" : 4,
                  "profile_medium" : "https://xxxxxx.cloudfront.net/pictures/athletes/123456789/123456789/2/medium.jpg",
                  "profile" : "https://xxxxx.cloudfront.net/pictures/athletes/123456789/123456789/2/large.jpg",
                  "friend" : null,
                  "follower" : null,
                  "follower_count" : 5,
                  "friend_count" : 5,
                  "mutual_friend_count" : 0,
                  "athlete_type" : 1,
                  "date_preference" : "%m/%d/%Y",
                  "measurement_preference" : "feet",
                  "clubs" : [ ],
                  "ftp" : null,
                  "weight" : 0,
                  "bikes" : [ {
                    "id" : "b12345678987655",
                    "primary" : true,
                    "name" : "EMC",
                    "resource_state" : 2,
                    "distance" : 0
                  } ],
                  "shoes" : [ {
                    "id" : "g12345678987655",
                    "primary" : true,
                    "name" : "adidas",
                    "resource_state" : 2,
                    "distance" : 4904
                  } ]
                }
              }
            }
          },
          "default" : {
            "description" : "Unexpected error.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Fault"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Athletes" ],
        "summary" : "Update Athlete",
        "description" : "Update the currently authenticated athlete. Requires profile:write scope.",
        "operationId" : "updateLoggedInAthlete",
        "parameters" : [ {
          "name" : "weight",
          "in" : "path",
          "description" : "The weight of the athlete in kilograms.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "number",
            "format" : "float"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Profile information for the authenticated athlete.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DetailedAthlete"
                },
                "example" : {
                  "id" : 12345678987655098765444,
                  "username" : "marianne_v",
                  "resource_state" : 3,
                  "firstname" : "Marianne",
                  "lastname" : "V.",
                  "city" : "San Francisco",
                  "state" : "CA",
                  "country" : "US",
                  "sex" : "F",
                  "premium" : true,
                  "created_at" : "2017-11-14T02:30:05Z",
                  "updated_at" : "2018-02-06T19:32:20Z",
                  "badge_type_id" : 4,
                  "profile_medium" : "https://xxxxxx.cloudfront.net/pictures/athletes/1234567898765509876/1234567898765509876/2/medium.jpg",
                  "profile" : "https://xxxxx.cloudfront.net/pictures/athletes/1234567898765509876/1234567898765509876/2/large.jpg",
                  "friend" : null,
                  "follower" : null,
                  "follower_count" : 5,
                  "friend_count" : 5,
                  "mutual_friend_count" : 0,
                  "athlete_type" : 1,
                  "date_preference" : "%m/%d/%Y",
                  "measurement_preference" : "feet",
                  "clubs" : [ ],
                  "ftp" : null,
                  "weight" : 0,
                  "bikes" : [ {
                    "id" : "b1234567898765509876",
                    "primary" : true,
                    "name" : "EMC",
                    "resource_state" : 2,
                    "distance" : 0
                  } ],
                  "shoes" : [ {
                    "id" : "g1234567898765509876",
                    "primary" : true,
                    "name" : "adidas",
                    "resource_state" : 2,
                    "distance" : 4904
                  } ]
                }
              }
            }
          },
          "default" : {
            "description" : "Unexpected error.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Fault"
                }
              }
            }
          }
        }
      }
    },
    "/athlete/zones" : {
      "get" : {
        "tags" : [ "Athletes" ],
        "summary" : "Get Zones",
        "description" : "Returns the the authenticated athlete's heart rate and power zones. Requires profile:read_all.",
        "operationId" : "getLoggedInAthleteZones",
        "responses" : {
          "200" : {
            "description" : "Heart rate and power zones.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Zones"
                },
                "example" : [ {
                  "distribution_buckets" : [ {
                    "max" : 0,
                    "min" : 0,
                    "time" : 1498
                  }, {
                    "max" : 50,
                    "min" : 0,
                    "time" : 62
                  }, {
                    "max" : 100,
                    "min" : 50,
                    "time" : 169
                  }, {
                    "max" : 150,
                    "min" : 100,
                    "time" : 536
                  }, {
                    "max" : 200,
                    "min" : 150,
                    "time" : 672
                  }, {
                    "max" : 250,
                    "min" : 200,
                    "time" : 821
                  }, {
                    "max" : 300,
                    "min" : 250,
                    "time" : 529
                  }, {
                    "max" : 350,
                    "min" : 300,
                    "time" : 251
                  }, {
                    "max" : 400,
                    "min" : 350,
                    "time" : 80
                  }, {
                    "max" : 450,
                    "min" : 400,
                    "time" : 81
                  }, {
                    "max" : -1,
                    "min" : 450,
                    "time" : 343
                  } ],
                  "type" : "power",
                  "resource_state" : 3,
                  "sensor_based" : true
                } ]
              }
            }
          },
          "default" : {
            "description" : "Unexpected error.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Fault"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "ActivityStats" : {
        "type" : "object",
        "properties" : {
          "biggest_ride_distance" : {
            "type" : "number",
            "description" : "The longest distance ridden by the athlete.",
            "format" : "double"
          },
          "biggest_climb_elevation_gain" : {
            "type" : "number",
            "description" : "The highest climb ridden by the athlete.",
            "format" : "double"
          },
          "recent_ride_totals" : {
            "$ref" : "#/components/schemas/ActivityTotal"
          },
          "recent_run_totals" : {
            "$ref" : "#/components/schemas/ActivityTotal"
          },
          "recent_swim_totals" : {
            "$ref" : "#/components/schemas/ActivityTotal"
          },
          "ytd_ride_totals" : {
            "$ref" : "#/components/schemas/ActivityTotal"
          },
          "ytd_run_totals" : {
            "$ref" : "#/components/schemas/ActivityTotal"
          },
          "ytd_swim_totals" : {
            "$ref" : "#/components/schemas/ActivityTotal"
          },
          "all_ride_totals" : {
            "$ref" : "#/components/schemas/ActivityTotal"
          },
          "all_run_totals" : {
            "$ref" : "#/components/schemas/ActivityTotal"
          },
          "all_swim_totals" : {
            "$ref" : "#/components/schemas/ActivityTotal"
          }
        },
        "description" : "A set of rolled-up statistics and totals for an athlete",
        "example" : {
          "recent_run_totals" : null,
          "all_run_totals" : null,
          "recent_swim_totals" : null,
          "biggest_ride_distance" : 0.8008281904610115,
          "ytd_swim_totals" : null,
          "all_swim_totals" : null,
          "recent_ride_totals" : {
            "distance" : 5.962134,
            "achievement_count" : 9,
            "count" : 1,
            "elapsed_time" : 2,
            "elevation_gain" : 7.0614014,
            "moving_time" : 5
          },
          "biggest_climb_elevation_gain" : 6.027456183070403,
          "ytd_ride_totals" : null,
          "all_ride_totals" : null,
          "ytd_run_totals" : null
        }
      },
      "ActivityTotal" : {
        "type" : "object",
        "properties" : {
          "count" : {
            "type" : "integer",
            "description" : "The number of activities considered in this total."
          },
          "distance" : {
            "type" : "number",
            "description" : "The total distance covered by the considered activities.",
            "format" : "float"
          },
          "moving_time" : {
            "type" : "integer",
            "description" : "The total moving time of the considered activities."
          },
          "elapsed_time" : {
            "type" : "integer",
            "description" : "The total elapsed time of the considered activities."
          },
          "elevation_gain" : {
            "type" : "number",
            "description" : "The total elevation gain of the considered activities.",
            "format" : "float"
          },
          "achievement_count" : {
            "type" : "integer",
            "description" : "The total number of achievements of the considered activities."
          }
        },
        "description" : "A roll-up of metrics pertaining to a set of activities. Values are in seconds and meters.",
        "example" : {
          "distance" : 5.962134,
          "achievement_count" : 9,
          "count" : 1,
          "elapsed_time" : 2,
          "elevation_gain" : 7.0614014,
          "moving_time" : 5
        }
      },
      "Fault" : {
        "type" : "object",
        "properties" : {
          "errors" : {
            "type" : "array",
            "description" : "The set of specific errors associated with this fault, if any.",
            "items" : {
              "$ref" : "#/components/schemas/Error"
            }
          },
          "message" : {
            "type" : "string",
            "description" : "The message of the fault."
          }
        },
        "description" : "Encapsulates the errors that may be returned from the API."
      },
      "Error" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string",
            "description" : "The code associated with this error."
          },
          "field" : {
            "type" : "string",
            "description" : "The specific field or aspect of the resource associated with this error."
          },
          "resource" : {
            "type" : "string",
            "description" : "The type of resource associated with this error."
          }
        }
      },
      "DetailedAthlete" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/SummaryAthlete"
        }, {
          "type" : "object",
          "properties" : {
            "follower_count" : {
              "type" : "integer",
              "description" : "The athlete's follower count."
            },
            "friend_count" : {
              "type" : "integer",
              "description" : "The athlete's friend count."
            },
            "measurement_preference" : {
              "type" : "string",
              "description" : "The athlete's preferred unit system.",
              "enum" : [ "feet", "meters" ]
            },
            "ftp" : {
              "type" : "integer",
              "description" : "The athlete's FTP (Functional Threshold Power)."
            },
            "weight" : {
              "type" : "number",
              "description" : "The athlete's weight.",
              "format" : "float"
            },
            "clubs" : {
              "type" : "array",
              "description" : "The athlete's clubs.",
              "items" : {
                "$ref" : "#/components/schemas/SummaryClub"
              }
            },
            "bikes" : {
              "type" : "array",
              "description" : "The athlete's bikes.",
              "items" : {
                "$ref" : "#/components/schemas/SummaryGear"
              }
            },
            "shoes" : {
              "type" : "array",
              "description" : "The athlete's shoes.",
              "items" : {
                "$ref" : "#/components/schemas/SummaryGear"
              }
            }
          }
        } ]
      },
      "SummaryAthlete" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/MetaAthlete"
        }, {
          "type" : "object",
          "properties" : {
            "resource_state" : {
              "type" : "integer",
              "description" : "Resource state, indicates level of detail. Possible values: 1 -> \"meta\", 2 -> \"summary\", 3 -> \"detail\""
            },
            "firstname" : {
              "type" : "string",
              "description" : "The athlete's first name."
            },
            "lastname" : {
              "type" : "string",
              "description" : "The athlete's last name."
            },
            "profile_medium" : {
              "type" : "string",
              "description" : "URL to a 62x62 pixel profile picture."
            },
            "profile" : {
              "type" : "string",
              "description" : "URL to a 124x124 pixel profile picture."
            },
            "city" : {
              "type" : "string",
              "description" : "The athlete's city."
            },
            "state" : {
              "type" : "string",
              "description" : "The athlete's state or geographical region."
            },
            "country" : {
              "type" : "string",
              "description" : "The athlete's country."
            },
            "sex" : {
              "type" : "string",
              "description" : "The athlete's sex.",
              "enum" : [ "M", "F" ]
            },
            "premium" : {
              "type" : "boolean",
              "description" : "Deprecated.  Use summit field instead. Whether the athlete has any Summit subscription."
            },
            "summit" : {
              "type" : "boolean",
              "description" : "Whether the athlete has any Summit subscription."
            },
            "created_at" : {
              "type" : "string",
              "description" : "The time at which the athlete was created.",
              "format" : "date-time"
            },
            "updated_at" : {
              "type" : "string",
              "description" : "The time at which the athlete was last updated.",
              "format" : "date-time"
            }
          }
        } ]
      },
      "MetaAthlete" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "description" : "The unique identifier of the athlete",
            "format" : "int64"
          }
        },
        "example" : {
          "id" : 1
        }
      },
      "SummaryClub" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/MetaClub"
        }, {
          "type" : "object",
          "properties" : {
            "profile_medium" : {
              "type" : "string",
              "description" : "URL to a 60x60 pixel profile picture."
            },
            "cover_photo" : {
              "type" : "string",
              "description" : "URL to a ~1185x580 pixel cover photo."
            },
            "cover_photo_small" : {
              "type" : "string",
              "description" : "URL to a ~360x176  pixel cover photo."
            },
            "sport_type" : {
              "type" : "string",
              "description" : "Deprecated. Prefer to use activity_types.",
              "enum" : [ "cycling", "running", "triathlon", "other" ]
            },
            "activity_types" : {
              "type" : "array",
              "description" : "The activity types that count for a club. This takes precedence over sport_type.",
              "items" : {
                "$ref" : "#/components/schemas/ActivityType"
              }
            },
            "city" : {
              "type" : "string",
              "description" : "The club's city."
            },
            "state" : {
              "type" : "string",
              "description" : "The club's state or geographical region."
            },
            "country" : {
              "type" : "string",
              "description" : "The club's country."
            },
            "private" : {
              "type" : "boolean",
              "description" : "Whether the club is private."
            },
            "member_count" : {
              "type" : "integer",
              "description" : "The club's member count."
            },
            "featured" : {
              "type" : "boolean",
              "description" : "Whether the club is featured or not."
            },
            "verified" : {
              "type" : "boolean",
              "description" : "Whether the club is verified or not."
            },
            "url" : {
              "type" : "string",
              "description" : "The club's vanity URL."
            }
          }
        } ]
      },
      "MetaClub" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "description" : "The club's unique identifier.",
            "format" : "int64"
          },
          "resource_state" : {
            "type" : "integer",
            "description" : "Resource state, indicates level of detail. Possible values: 1 -> \"meta\", 2 -> \"summary\", 3 -> \"detail\""
          },
          "name" : {
            "type" : "string",
            "description" : "The club's name."
          }
        }
      },
      "ActivityType" : {
        "type" : "string",
        "description" : "An enumeration of the types an activity may have.",
        "enum" : [ "AlpineSki", "BackcountrySki", "Canoeing", "Crossfit", "EBikeRide", "Elliptical", "Golf", "Handcycle", "Hike", "IceSkate", "InlineSkate", "Kayaking", "Kitesurf", "NordicSki", "Ride", "RockClimbing", "RollerSki", "Rowing", "Run", "Sail", "Skateboard", "Snowboard", "Snowshoe", "Soccer", "StairStepper", "StandUpPaddling", "Surfing", "Swim", "Velomobile", "VirtualRide", "VirtualRun", "Walk", "WeightTraining", "Wheelchair", "Windsurf", "Workout", "Yoga" ]
      },
      "SummaryGear" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "description" : "The gear's unique identifier."
          },
          "resource_state" : {
            "type" : "integer",
            "description" : "Resource state, indicates level of detail. Possible values: 2 -> \"summary\", 3 -> \"detail\""
          },
          "primary" : {
            "type" : "boolean",
            "description" : "Whether this gear's is the owner's default one."
          },
          "name" : {
            "type" : "string",
            "description" : "The gear's name."
          },
          "distance" : {
            "type" : "number",
            "description" : "The distance logged with this gear.",
            "format" : "float"
          }
        }
      },
      "Zones" : {
        "type" : "object",
        "properties" : {
          "heart_rate" : {
            "$ref" : "#/components/schemas/HeartRateZoneRanges"
          },
          "power" : {
            "$ref" : "#/components/schemas/PowerZoneRanges"
          }
        },
        "example" : {
          "heart_rate" : {
            "custom_zones" : true,
            "zones" : [ {
              "min" : 0,
              "max" : 6
            }, {
              "min" : 0,
              "max" : 6
            } ]
          },
          "power" : {
            "zones" : null
          }
        }
      },
      "HeartRateZoneRanges" : {
        "type" : "object",
        "properties" : {
          "custom_zones" : {
            "type" : "boolean",
            "description" : "Whether the athlete has set their own custom heart rate zones"
          },
          "zones" : {
            "$ref" : "#/components/schemas/ZoneRanges"
          }
        },
        "example" : {
          "custom_zones" : true,
          "zones" : [ {
            "min" : 0,
            "max" : 6
          }, {
            "min" : 0,
            "max" : 6
          } ]
        }
      },
      "ZoneRanges" : {
        "type" : "array",
        "items" : {
          "$ref" : "#/components/schemas/ZoneRange"
        }
      },
      "ZoneRange" : {
        "type" : "object",
        "properties" : {
          "min" : {
            "type" : "integer",
            "description" : "The minimum value in the range."
          },
          "max" : {
            "type" : "integer",
            "description" : "The maximum value in the range."
          }
        },
        "example" : {
          "min" : 0,
          "max" : 6
        }
      },
      "PowerZoneRanges" : {
        "type" : "object",
        "properties" : {
          "zones" : {
            "$ref" : "#/components/schemas/ZoneRanges"
          }
        },
        "example" : {
          "zones" : null
        }
      },
...
    },
...
    "securitySchemes" : {
      "strava_oauth" : {
        "type" : "oauth2",
        "flows" : {
          "authorizationCode" : {
            "authorizationUrl" : "https://www.strava.com/api/v3/oauth/authorize",
            "tokenUrl" : "https://www.strava.com/api/v3/oauth/token",
            "scopes" : {
              "read" : "Read public segments, public routes, public profile data, public posts, public events, club feeds, and leaderboards",
              "read_all" : "Read private routes, private segments, and private events for the user",
              "profile:read_all" : "Read all profile information even if the user has set their profile visibility to Followers or Only You",
              "profile:write" : "Update the user's weight and Functional Threshold Power (FTP), and access to star or unstar segments on their behalf",
              "activity:read" : "Read the user's activity data for activities that are visible to Everyone and Followers, excluding privacy zone data",
              "activity:read_all" : "The same access as activity:read, plus privacy zone data and access to read the user's activities with visibility set to Only You",
              "activity:write" : "Access to create manual activities and uploads, and access to edit any activities that are visible to the app, based on activity read access level"
            }
          }
        }
      }
    }
  },
...
}